refactor: 액세스 토큰 만료 시 actor가 재발급을 처리해 토큰 일관성 보장
